Allow solveable substitutions
Summary:
This extends the current valid substitution functionality to allow substitutions which involve unsolved constraints to appear in the list of valid substitutions. This allows e.g. substitutions like `maxmimum` for holes of type `Ord a => [a] -> a`, even though the constraint solver has an unsolved pending superclass constraint for the `Ord a`.

This also filters out suggestions from the `GHC.Err` module, since even though `error` and `undefined` are indeed valid substitutions, they are "trivial", and almost never useful to the user.

Reviewers: bgamari

Subscribers: rwbarton, thomie, carter

Differential Revision: https://phabricator.haskell.org/D4444