Add sortBySize sorting for valid substitutions and use as the default
The previous default where we sort by subsumption could be very slow,
especially if there were a lot of valid substitutions. The sort by size
sorts the valid substitutions not by subsumption but by the size of the
types needed to turn the valid substitution into the type of the hole.
This takes a lot less work, but still gives reasonable results in most
cases. The original subsumption sorting algorithm is still available if
users are willing to pay the performance penalty, using the flag
`-fsort-by-subsumption-substitutions`.