Fix errors caused by invalid candidates leaking from hole fits
Summary:
This is a one line fix (and a note) that fixes three tickets, #15007, #15321 and #15202.

The issue was that errors caused by illegal candidates (according to GHC stage or being internal names) were leaking to the user, causing bewildering error messages. If a candidate causes the type checker to error, it is not a valid hole fit, and should be discarded.

As mentioned in #15321, this can cause a pattern of omissions, which might be hard to discover. A better approach would be to gather the error messages, and ask users to report them as GHC bugs. This will be implemented in a subsequent change.

Reviewers: bgamari

Subscribers: rwbarton, thomie, carter

GHC Trac Issues: #15007, #15321, #15202

Differential Revision: https://phabricator.haskell.org/D4909